Use a 64 bit integer to store total perft nodes
This was overflowing into the sign bit past perft 6 from the startpos.

+++ src/uci.cc
@@ -86,7 +86,7 @@ int Uci::_perft(const Board& board, int depth) {
 }
 
 void Uci::_perftDivide(int depth) {
-  int total = 0;
+  unsigned long long total = 0;
 
   MoveGen movegen(_board);
